requirements.txt
requests
discord.py
configparser
python-dotenv
pytz